Edit/Delete photo from feed and comments
Fix selecting fixed photo list

Implement bucket item edit/delete

+++ app/src/main/java/com/worldventures/dreamtrips/modules/tripsimages/view/fragment/FullScreenPhotoFragment.java
@@ -1,5 +1,6 @@
 package com.worldventures.dreamtrips.modules.tripsimages.view.fragment;
 
+import android.app.Dialog;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Build;
@@ -24,9 +25,13 @@ import com.facebook.drawee.view.SimpleDraweeView;
 import com.facebook.imagepipeline.request.ImageRequest;
 import com.techery.spares.annotations.Layout;
 import com.worldventures.dreamtrips.R;
+import com.worldventures.dreamtrips.core.navigation.NavigationBuilder;
+import com.worldventures.dreamtrips.core.navigation.Route;
+import com.worldventures.dreamtrips.core.navigation.ToolbarConfig;
 import com.worldventures.dreamtrips.core.utils.ViewUtils;
 import com.worldventures.dreamtrips.modules.common.view.fragment.BaseFragmentWithArgs;
 import com.worldventures.dreamtrips.modules.common.view.util.TextWatcherAdapter;
+import com.worldventures.dreamtrips.modules.tripsimages.bundle.EditPhotoBundle;
 import com.worldventures.dreamtrips.modules.tripsimages.bundle.FullScreenPhotoBundle;
 import com.worldventures.dreamtrips.modules.tripsimages.model.Flag;
 import com.worldventures.dreamtrips.modules.tripsimages.model.IFullScreenObject;
@@ -73,8 +78,10 @@ public class FullScreenPhotoFragment<T extends IFullScreenObject>
     protected ImageView ivShare;
     @InjectView(R.id.iv_flag)
     protected ImageView ivFlag;
-    @InjectView(R.id.iv_delete)
-    protected ImageView ivDelete;
+    @InjectView(R.id.more)
+    protected ImageView more;
+    @InjectView(R.id.delete)
+    protected ImageView delete;
     @InjectView(R.id.user_photo)
     protected SimpleDraweeView civUserPhoto;
     @InjectView(R.id.checkBox)
@@ -219,24 +226,51 @@ public class FullScreenPhotoFragment<T extends IFullScreenObject>
         }
     }
 
-    @OnClick(R.id.iv_delete)
-    public void actionDelete() {
-        new MaterialDialog.Builder(getActivity())
-                .title(R.string.delete_photo_title)
-                .content(R.string.delete_photo_text)
-                .positiveText(R.string.delete_photo_positiove)
-                .negativeText(R.string.delete_photo_negative)
-                .callback(new MaterialDialog.ButtonCallback() {
-                    @Override
-                    public void onPositive(MaterialDialog dialog) {
-                        getPresenter().onDeleteAction();
-                    }
+    @OnClick(R.id.more)
+    public void actionMore() {
+        PopupMenu popup = new PopupMenu(getContext(), more);
+        popup.inflate(R.menu.menu_photo_edit);
+        popup.setOnMenuItemClickListener(item -> {
+            switch (item.getItemId()) {
+                case R.id.action_delete:
+                    deletePhoto();
+                    break;
+                case R.id.action_edit:
+                    getPresenter().onEdit();
+                    break;
+            }
 
-                    @Override
-                    public void onNegative(MaterialDialog dialog) {
-                        dialog.dismiss();
-                    }
-                }).show();
+            return true;
+        });
+        popup.show();
+    }
+
+    @OnClick(R.id.delete)
+    public void delete() {
+        deletePhoto();
+    }
+
+    private void deletePhoto() {
+        Dialog dialog = new SweetAlertDialog(getContext(), SweetAlertDialog.WARNING_TYPE)
+                .setTitleText(getResources().getString(R.string.photo_delete))
+                .setContentText(getResources().getString(R.string.photo_delete_caption))
+                .setConfirmText(getResources().getString(R.string.post_delete_confirm))
+                .setConfirmClickListener(sDialog -> {
+                    sDialog.dismissWithAnimation();
+                    getPresenter().onDeleteAction();
+                });
+        dialog.setCanceledOnTouchOutside(true);
+        dialog.show();
+    }
+
+    @Override
+    public void setMoreVisibility(boolean visible) {
+        if (visible) {
+            more.setVisibility(View.VISIBLE);
+        } else {
+            more.setVisibility(View.GONE);
+        }
+
     }
 
     @OnClick(R.id.iv_like)
@@ -253,6 +287,14 @@ public class FullScreenPhotoFragment<T extends IFullScreenObject>
     public void actionComment() {
     }
 
+    @Override
+    public void openEdit(EditPhotoBundle bundle) {
+        NavigationBuilder.create()
+                .with(activityRouter)
+                .toolbarConfig(ToolbarConfig.Builder.create().visible(false).build())
+                .data(bundle)
+                .attach(Route.PHOTO_EDIT);
+    }
 
     @Override
     public void setFlags(List<Flag> flags) {
@@ -418,9 +460,9 @@ public class FullScreenPhotoFragment<T extends IFullScreenObject>
     @Override
     public void setDeleteVisibility(boolean isVisible) {
         if (isVisible) {
-            ivDelete.setVisibility(View.VISIBLE);
+            delete.setVisibility(View.VISIBLE);
         } else {
-            ivDelete.setVisibility(View.GONE);
+            delete.setVisibility(View.GONE);
         }
     }
